qa: remove PHPUnit warnings
We've had a number of warnings from PHPUnit about usage of deprecated assertions.
However, in many cases, there is no easy way to test that attributes were set other than reflection.
This patch adds an `AttributeAssertionTrait` that recreates the ones we use, and is now composed in test cases where the assertions are used.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>