feat: event-driven handler runner
This patch builds on #24 to expand the event dispatcher paradigm such that the `SwooleRequestHandlerRunner` simply dispatches events via a PSR-14 dispatcher.
All previous method bodies have been extracted to listener classes, and each now creates a typed event that composes the arguments passed via the Swoole dispatcher.
We can have conditional functionality now by providing listeners, which are defined in a nested array in the configuration, mapping to dedicated container services.

The functionality requires testing, and should likely be expanded to additional events (tasks come to mind).
It may even serve as a basis for providing a Promise implementation.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>