fix: ensure `onTask()` works correctly when task_enable_coroutine is enabled
Per @basz, and verified locally, when `task_enable_coroutine` is enabled, the task callback receives an object as the second and last argument.
That object then contains the task id, worker id, and data as properties, and has its own `finish()` method that should be called instead of the server's when complete.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>