fix: ensure AccessLogDataMap does not throw fatal error due to uninitialized typed properties
The `AccessLogDataMap` has named constructors that ensure either the `$psrResponse` or `$staticResource` properties are initialized.
However, internally, it also checks to see if one or the other is _set_ in order to determine which actions to take and/or values to return.
When uninitialized, these can lead to fatal errors.

However, when `null`, we get `PossiblyNullReference` notices from Psalm.
To fix these, we need to do proper conditionals instead of ternary statements, and raise exceptions when neither are present.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>