feat: allow access to dispatcher and Swoole HTTP server in extension classes
To allow changing how events are attached, particularly when using third-party extensions that need to attach directly to the server and decorate the main functionality (e.g., timers), the swoole instance should be exposed.
Exposing the dispatcher also allows extending the various internal methods in a way that allows their individual logic to be wrapped.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>