fix: autoconvert strftime formats to ICU, and use IntlDateFormatter for access log request representations
AccessLogFormatter calls on AccesLogDataMap to format a number of strings, including request times; these are generally used in cookies and/or header values.
Previously, we used `strftime()` to format date/time values.
However, this is deprecated starting in PHP 8.1, with the recommendation of using `IntlDateFormatter` for purposes of formatting date/time strings in a locale-specific way.
Because I'm unclear on whether the formats will always come from internal classes, or if they might come from external sources (e.g., cookie values), I've created a new final class, `Mezzio\Swoole\StrftimeToICUFormatMap` for mapping strftime formats to the ICU equivalents.
I've added some basic tests for the class, which include demonstrating that the formats we use internally all get translated correctly.
I've also added a test case for `AccessLogDataMap::getRequestTime()` to demonstrate that it generates the same values.

Targeting the 3.5 series, as this is technically a bug in our PHP 8.1 support.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>