fix: pass DateTimeImmutable instead of calling time()
The StrftimeToICUFormatMap case for `%s` was calling `return time()`, which (a) was the incorrect type (a string should be returned), and (b) could lead to a mismatch between the request time and what is reported in the logs.

StrftimeToICUFormatMap::mapStrftimeToICU now accepts a second argument, a DateTimeInterface $requestTime, and this is used when `%s` is in the format.
AccessLogDataMap::getRequestTime creates a DateTimeImmutable instance from the server $request_time_float property and passes that value to StrftimeToICUFormatMap.

Creation of the callback for preg_replace_callback has been moved to another method, based on issues reported by Psalm; the callback now has full typehinting as well.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>